.github/translation/requirements.txt
openai>=1.40.0
deepl==1.18.0
pyyaml
